requirements.txt
langchain
sentence_transformers
openai
pinecone-client
unstructured[all-docs]
elevenlabs
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-speech
pyaudio
plivo
ngrok


'''
Other packages to be installed:
1. google cloud sdk
3. ngrok
'''